Enhance app renewal functionality in contracts
- Updated the `DeployAppAccount` library to reserve space for an additional selector related to app renewal.
- Added `renewApp` function in `AppRegistryFacet` to facilitate app renewal, including parameters for app address, account, and data.
- Extended `IAppRegistry` interface to include the new `renewApp` method, ensuring external calls can handle app renewals effectively.

+++ packages/contracts/src/spaces/facets/account/AppAccountBase.sol
@@ -140,7 +140,7 @@ abstract contract AppAccountBase is
         if (app.appId == EMPTY_UID) AppNotRegistered.selector.revertWith();
 
         // Calculate the new expiration time (extends current expiration by app duration)
-        uint64 newExpiration = _getAppExpiration(app.appId, app.duration);
+        uint48 newExpiration = _getAppExpiration(app.appId, app.duration);
 
         // Update the group expiration
         _setGroupExpiration(app.appId, newExpiration);
